Summary: Fixed grayed-out button issue. Stability improvements.
Details: JPAdvert would gray out (ad.enabled = NO) when there was no available link to visit; now it is not grayed out it instead the target/action pair is removed letting it act as a static plain view. This is accomplished by checking when loading the adURL, check to see if it's `nil` or empty ("").
Added ()s to ternary conditional operator.